% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsvs2_temp_3.R
\name{bsvs2_temp}
\alias{bsvs2_temp}
\title{Bayesian Iterated Screening (ultra-high, high or low dimensional).}
\usage{
bsvs2_temp(xmat, ys, xty, lam, w, k, D, xbar, temp, temp.multip, logp.best,
  r.idx.best, n, ncovar)
}
\description{
Perform Bayesian iterated screening in Gaussian regression models
}
