% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{bwas}
\alias{bwas}
\title{BWAS: Bayesian GWAS for a Single Trait}
\usage{
bwas(x, y, params, bigx, ehits = 20)
}
\arguments{
\item{x}{Cleaned SNP matrix.}

\item{y}{Phenotype vector.}

\item{params}{Named numeric vector with lambda and w.}

\item{bigx}{Full SNP matrix.}

\item{ehits}{Expected number of hits (default 20).}
}
\description{
Runs SVEN and UNITE for one trait.
}
