% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{calc.runtime}
\alias{calc.runtime}
\title{Estimate SVEN Runtime}
\usage{
calc.runtime(X)
}
\arguments{
\item{X}{SNP matrix.}
}
\description{
Times a single SVEN run on the given SNP matrix using a random phenotype.
}
