% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{clean}
\alias{clean}
\title{Clean SNP Matrix}
\usage{
clean(SNPmat, MAF_threshold = 0.05)
}
\arguments{
\item{SNPmat}{A sparse SNP matrix (dgCMatrix).}

\item{MAF_threshold}{Minor Allele Frequency cutoff (default 0.05).}
}
\description{
Removes duplicate SNPs and filters low MAF variants.
}
