% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{create_param_mat}
\alias{create_param_mat}
\title{Create Parameter Grid}
\usage{
create_param_mat(x)
}
\arguments{
\item{x}{Cleaned SNP matrix.}
}
\description{
Builds a grid of lambda and w tuning parameters for SVEN.
}
