% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dense2sparse.R
\name{dense2sparse}
\alias{dense2sparse}
\title{Convert numeric genotype matrix to sparse matrix}
\usage{
dense2sparse(
  file.name,
  num.genotypes = NULL,
  separator = NULL,
  fixed = TRUE,
  recode = TRUE,
  na.strings = c("NA", ".", "-9", "NN"),
  progress = TRUE
)
}
\arguments{
\item{file.name}{Path to the numeric genotype file. Could be (and should be) gzipped.}

\item{num.genotypes}{Maximum number of genotypes to read. An upper bound is OK.}

\item{separator}{"\\t" or "," etc. that separates the entries in a line.}

\item{fixed}{Is `separator` a fixed string (TRUE) or a regex (FALSE)?}

\item{recode}{recode each SNP so its major allele is 0 (per-SNP)}

\item{na.strings}{tokens treated as missing}

\item{progress}{Whether to show a progress bar (default TRUE).}
}
\value{
A sparse matrix of class dgCMatix.
}
\description{
Reads a numeric genotype file and converts it to a sparse matrix format.
}
