% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_convert.R
\name{dense_to_sparse_converter}
\alias{dense_to_sparse_converter}
\title{Launch the Sparse Converter Shiny App}
\usage{
dense_to_sparse_converter()
}
\description{
Opens the Sparse Matrix Converter GUI in your browser.
}
