% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{FDR_corrected}
\alias{FDR_corrected}
\title{FDR using correlation threshold}
\usage{
FDR_corrected(model, truth, x, threshold = 0.9)
}
\arguments{
\item{model}{Integer vector of selected SNP indices.}

\item{truth}{Integer vector of true causal SNP indices.}

\item{x}{SNP matrix.}

\item{threshold}{Correlation threshold (default 0.9).}
}
\description{
Computes False Discovery Rate using SNP correlation as proximity measure.
}
