% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{jcidx}
\alias{jcidx}
\title{Jaccard Index}
\usage{
jcidx(vars, truth)
}
\arguments{
\item{vars}{Integer vector of selected SNP indices.}

\item{truth}{Integer vector of true causal SNP indices.}
}
\description{
Computes Jaccard index between selected and true causal SNPs.
}
