% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybsvs_hd2.R
\name{mybsvs}
\alias{mybsvs}
\title{Bayesian Variable Selection (ultra-high, high or low dimensional).}
\usage{
mybsvs(X, y, w, lam, temp.multip = 3, M = 200, trueidx)
}
\arguments{
\item{X}{An \eqn{n x p} matrix. Sparse matrices are supported and every
care is taken not to make copies of this (typically) giant matrix.
No need to center or scale.}

\item{y}{The response vector of length \code{n}.}

\item{w}{The prior inclusion probability of each variable.}

\item{lam}{The slab precision parameter.
as suggested by the theory of Wang et al. (2019).}

\item{temp.multip}{The temperature multiple. Default: 3.}

\item{M}{The number of iteration. Default: 200.}
}
\description{
Perform Bayesian Variable Selection in Gaussian regression models
}
