% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{parameter_selection}
\alias{parameter_selection}
\title{Select Optimal Tuning Parameters}
\usage{
parameter_selection(
  X,
  R2 = 0.5,
  betamax = 1,
  n.cores = max(1, parallel::detectCores() - 1),
  hitsize = "all",
  MAF_threshold = 0.05
)
}
\arguments{
\item{X}{Raw SNP matrix (dgCMatrix).}

\item{R2}{Heritability (default 0.5).}

\item{betamax}{Maximum effect size magnitude (default 1).}

\item{n.cores}{Number of cores (default: detectCores() - 1).}

\item{hitsize}{One of "all", "small", "medium", "large" (default "all").}

\item{MAF_threshold}{MAF cutoff (default 0.05).}
}
\description{
Full training step: cleans SNP matrix and tunes SVEN hyperparameters.
}
