% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bsvs.R
\name{predict.bsvs}
\alias{predict.bsvs}
\title{make predictions from a "bsvs" object.}
\usage{
\method{predict}{bsvs}(bsvs.obj, newx, model = "MAP")
}
\arguments{
\item{bsvs.obj}{fitted bsvs model object}

\item{newx}{Matrix of new values for x at which predictions are to be made. Must be a
matrix; can be sparse as in Matrix package.}

\item{model}{The model to be used to make predictions. Model "MAP" gives the fitted
values calculated using coefficients estimated from the MAP model; model "WAM" gives
the predicted values calculated using coefficients estimated from the WAM model. Default is "MAP".}
}
\value{
a vector of fitted values
}
\description{
Similar to other predict methods, this functions predicts fitted
values from a fitted "bsvs" object.
}
