% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sven.R
\name{sven}
\alias{sven}
\title{Bayesian Variable Selection in Gaussian regression models
(ultra-high, high or low dimensional).}
\usage{
sven(
  X,
  y,
  w = sqrt(nrow(X))/ncol(X),
  lam = nrow(X)/ncol(X)^2,
  Ntemp = 3,
  Tmax = (log(log(ncol(X))) + log(ncol(X))),
  Miter = 50,
  wam.threshold = 0.5,
  log.eps = -16
)
}
\arguments{
\item{X}{The \eqn{\code{n} \times \code{p}} covariate matrix. Sparse matrices
are supported and every care is taken not to make copies of this (typically)
giant matrix. No need to center or scale.}

\item{y}{The response vector of length \code{n}.}

\item{w}{The prior inclusion probability of each variable. Default: \code{sqrt(n)/p}.}

\item{lam}{The slab precision parameter. Default: \code{n/p^2}
as suggested by the theory of Wang et al. (2019).}

\item{Ntemp}{The number of temperatures. Default: 3.}

\item{Tmax}{The maximum temperature. Default: log(log(\code{p}))+log(\code{p}).}

\item{Miter}{The number of iteration. Default: 50.}

\item{wam.threshold}{The threshold probablity to select the covariates for WAM.
A covariate will be included in WAM if its corresponding marginal inclusion
probability is greater than the threshold. Default: 0.5.}

\item{log.eps}{The tolerance Default: -16.}
}
\value{
A list with components
\item{model.map}{A vector of indices corresponding to the selected variables
in the MAP model.}
\item{model.wam}{A vector of indices corresponding to the selected variables
in the WAM.}
\item{model.top}{A sparse matrix storing the top models.}
\item{beta.map}{The ridge estimator of regression coefficients in the MAP model.}
\item{beta.wam}{The ridge estimator of regression coefficients in the WAM.}
\item{mip.map}{The marginal inclusion probabilities of the variables in the MAP model.}
\item{mip.wam}{The marginal inclusion probabilities of the variables in the WAM.}
\item{pprob.map}{The log (unnormalized) posterior probability corresponding
to the MAP model.}
\item{pprob.top}{A vector of the log (unnormalized) posterior probabilities
corresponding to the top models.}
\item{stats}{Additional sstatistics.}
}
\description{
This function performs Bayesian Variable Selection in Gaussian
regression models
}
\examples{
n = 400; p = 20000; nonzero = 15
trueidx <- 1:15
nonzero.value <- 3
TrueBeta = numeric(ncovar)
TrueBeta[trueidx] <- nonzero.value

# Generate a matrix with group structure
z0 <- matrix(rnorm(n*(p-nonzero)), n, p-nonzero)
zz = matrix(rnorm(3*n), n, 3)
z = cbind(matrix(rep(zz, nonzero/3), nrow = n), z0)
e = cbind(matrix(rnorm(n*nonzero, mean = 0, sd = 0.1), nrow = n), matrix(0, n, p-nonzero))
X = z + e
y <- 0.5 + X \%*\% TrueBeta + rnorm(n)
res <- sven(X = X, y = y, lam = 200, w = 0.02, Ntemp = 9, Miter = 200)
res$model.map # the MAP model
res$model.wam # the WAM
res$mip.map # the marginal inclusion probabilities of the variables in the MAP model
res$mip.wam # the marginal inclusion probabilities of the variables in the WAM
res$pprob.top # the log (unnormalized) posterior probabilities corresponding to the top models.

res$beta.map # the ridge estimator of regression coefficients in the MAP model 
res$beta.wam # the ridge estimator of regression coefficients in the WAM
}
\references{

}
\author{
Dongjin Li and Somak Dutta\cr Maintainer:
Dongjin Li <dongjl@iastate.edu>
}
