% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{tune.sven.all}
\alias{tune.sven.all}
\title{Tune SVEN for All Hit Sizes}
\usage{
tune.sven.all(x, R2 = 0.5, betamax = 1, n.cores)
}
\arguments{
\item{x}{Cleaned SNP matrix.}

\item{R2}{Heritability (default 0.5).}

\item{betamax}{Maximum effect size magnitude (default 1).}

\item{n.cores}{Number of cores for parallel computation.}
}
\description{
Tunes SVEN parameters for small, medium, and large expected hit sizes.
}
