% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{tune.sven}
\alias{tune.sven}
\title{Tune SVEN Parameters}
\usage{
tune.sven(x, R2 = 0.5, ehits = 20, betamax = 1, n.cores)
}
\arguments{
\item{x}{Cleaned SNP matrix.}

\item{R2}{Heritability (default 0.5).}

\item{ehits}{Expected number of causal SNPs.}

\item{betamax}{Maximum effect size magnitude (default 1).}

\item{n.cores}{Number of cores for parallel computation.}
}
\description{
Runs 100 parallel simulations and returns the optimal (lambda, w) pair.
}
