% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{unite.sven}
\alias{unite.sven}
\title{UNITE: Post-process SVEN Model}
\usage{
unite.sven(x, bigx, basic_sven_object, y, ehits, threshold = 0)
}
\arguments{
\item{x}{Cleaned SNP matrix.}

\item{bigx}{Full (uncleaned) SNP matrix.}

\item{basic_sven_object}{Fitted SVEN object.}

\item{y}{Phenotype vector.}

\item{ehits}{Expected number of hits.}

\item{threshold}{MIP threshold (default 0).}
}
\description{
Propagates MIPs from SVEN hits to the full SNP matrix via LD.
}
