% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_functions.R
\name{dense2sparse}
\alias{dense2sparse}
\title{Convert numeric matrix to sparse matrix}
\usage{
dense2sparse(file.name, num.genotypes, separator, progress = TRUE)
}
\arguments{
\item{file.name}{Path to the numeric genotype file. Could be (and should be) gzipped.}

\item{num.genotypes}{Maximum number of genotypes to read. An upper bound is OK.}

\item{separator}{"\\t" or "," etc that separates the entries in a line.}

\item{progress}{Whether to show a progress bar (default TRUE).}
}
\description{
Reads a numeric genotype file and converts it to a sparse matrix format.
}
