% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sven.R
\name{sven}
\alias{sven}
\title{Selection of variables with embedded screening using Bayesian methods (SVEN) 
in Gaussian linear models (ultra-high, high or low dimensional).}
\usage{
sven(
  X,
  y,
  w = sqrt(nrow(X))/ncol(X),
  lam = nrow(X)/ncol(X)^2,
  Ntemp = 3,
  Tmax = (log(log(ncol(X))) + log(ncol(X))),
  Miter = 50,
  wam.threshold = 0.5,
  log.eps = -16
)
}
\arguments{
\item{X}{The \eqn{\code{n} \times \code{p}} covariate matrix. Sparse matrices
are supported and every care is taken not to make copies of this (typically)
giant matrix. No need to center or scale.}

\item{y}{The response vector of length \code{n}.}

\item{w}{The prior inclusion probability of each variable. Default: \code{sqrt(n)/p}.}

\item{lam}{The slab precision parameter. Default: \code{n/p^2}
as suggested by the theory of Wang et al. (2019).}

\item{Ntemp}{The number of temperatures. Default: 3.}

\item{Tmax}{The maximum temperature. Default: log(log(\code{p}))+log(\code{p}).}

\item{Miter}{The number of iteration. Default: 50.}

\item{wam.threshold}{The threshold probablity to select the covariates for WAM.
A covariate will be included in WAM if its corresponding marginal inclusion
probability is greater than the threshold. Default: 0.5.}

\item{log.eps}{The tolerance to choose the number of top models. Default: -16.}
}
\value{
A list with components
\item{model.map}{A vector of indices corresponding to the selected variables
in the MAP model.}
\item{model.wam}{A vector of indices corresponding to the selected variables
in the WAM.}
\item{model.top}{A sparse matrix storing the top models.}
\item{beta.map}{The ridge estimator of regression coefficients in the MAP model.}
\item{beta.wam}{The ridge estimator of regression coefficients in the WAM.}
\item{mip.map}{The marginal inclusion probabilities of the variables in the MAP model.}
\item{mip.wam}{The marginal inclusion probabilities of the variables in the WAM.}
\item{pprob.map}{The log (unnormalized) posterior probability corresponding
to the MAP model.}
\item{pprob.top}{A vector of the log (unnormalized) posterior probabilities
corresponding to the top models.}
\item{stats}{Additional statistics.}
}
\description{
SVEN is an approach to selecting variables with embedded screening 
using Bayesian method. It is also a variable selection method in the spirit of 
the stochastic shotgun search algorithm. However, by embedding a unique model 
based screening and using fast Cholesky updates, SVEN produces a highly scalable 
algorithm to explore gigantic model spaces and rapidly identify the regions of 
high posterior probabilities. It outputs the log (unnormalized) posterior 
probability of a set of best (highest probability) models, as well as the marginal 
inclusion probabilities corresponding to those models. For more details, see ...
}
\details{
SVEN is developed based on a hierarchical Gaussian linear model with priors placed 
on the regression coefficients as well as on the model space. Degenerate spike priors 
on inactive variables and Gaussian slab priors on active covariates makes the posterior 
probability (up to a normalizing constant) of a model \eqn{P(\gamma|Y)} available in 
explicit form. The variable selection starts from an empty model and update the model 
according to the posterior probability of its neighboring models for some pre-specified 
number of iterations. In each iteration, the models with large probabilities are screened 
out in order to quickly identify the regions of high posterior probabilities. A temperature 
schedule is used in order to recover models with large posterior probabilities and mitigate 
posterior multimodality associated with variable selection models. The default maximum 
temperature is guided by the posterior model selection consistency asymptotics in (...).

SVEN provides the maximum a posteriori (MAP) model as well as the weighted average model 
(WAM). WAM is obtained in the following way: (1) keep the best (highest probability) \eqn{K} 
models where \eqn{K} is chosen so that \eqn{P(\gamma^{(K)}|y)/P(\gamma^{(1)}|y) > \code{log.eps}};
(2) assign the weights \eqn{w_i = P(\gamma^{(i)}|y)/\sum_{k=1}^K P(\gamma^{(k)}|y)}
to the model \eqn{\gamma^{(i)}}; (3) define the approximate marginal inclusion probabilities 
for the \eqn{j}th variable as \eqn{\hat\pi_j = \sum_{k=1}^K w_k I(\gamma^{(k)}_j = 1)}. 
Then, the WAM is defined as the model containing variables \eqn{j} with 
\eqn{\hat\pi_j > \code{wam.threshold}}. SVEN also provides all the top \eqn{K} models which
are stored in an \eqn{p \times K} sparse matrix, along with their corresponding log (unnormalized) 
posterior probabilities.
}
\examples{
n = 400; p = 20000; nonzero = 15
trueidx <- 1:15
nonzero.value <- 3
TrueBeta = numeric(ncovar)
TrueBeta[trueidx] <- nonzero.value

# Generate a matrix with group structure
z0 <- matrix(rnorm(n*(p-nonzero)), n, p-nonzero)
zz = matrix(rnorm(3*n), n, 3)
z = cbind(matrix(rep(zz, nonzero/3), nrow = n), z0)
e = cbind(matrix(rnorm(n*nonzero, mean = 0, sd = 0.1), nrow = n), matrix(0, n, p-nonzero))
X = z + e
y <- 0.5 + X \%*\% TrueBeta + rnorm(n)
res <- sven(X = X, y = y, lam = 200, w = 0.02, Ntemp = 9, Miter = 200)
res$model.map # the MAP model
res$model.wam # the WAM
res$mip.map # the marginal inclusion probabilities of the variables in the MAP model
res$mip.wam # the marginal inclusion probabilities of the variables in the WAM
res$pprob.top # the log (unnormalized) posterior probabilities corresponding to the top models.

res$beta.map # the ridge estimator of regression coefficients in the MAP model 
res$beta.wam # the ridge estimator of regression coefficients in the WAM
}
\references{

}
\author{
Dongjin Li and Somak Dutta\cr Maintainer:
Dongjin Li <dongjl@iastate.edu>
}
